backend/requirements.txt
# backend/requirements.txt

fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-openai
langgraph
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests